Backend-Python/semantic_matching/requirements.txt
streamlit
sentence-transformers
faiss-cpu
rapidfuzz
spacy
numpy
pdfplumber
# Add these to existing requirements
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai>=0.8.0
google-cloud-speech>=2.26.0
google-cloud-texttospeech>=2.16.0
